backend/requirements.txt
flask
flask-cors
transformers
torch
sentencepiece
gunicorn
python-dotenv
